// admin-clients/libs/cpanel/venues/feature/standard-venue-tpl-manager/src/lib/graphic-view/model/svf-defs.enum.ts
export enum SVGDefs {
    //elements
    seatElement = 'circle',
    //properties
    idProperty = 'id',
    typeProperty = 'obType',
    //Classes
    selectionRectClass = 'selected-rect',
    filteredClass = 'filtered',
    interactiveClass = 'interactive',
    fitSvgAnimationClass = 'fit-svg-animation',
    venueTemplateSvgClass = 'venue-template-svg'
}